% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfitting.R
\name{Ramanpk}
\alias{Ramanpk}
\title{Raman peak fit}
\usage{
Ramanpk(data.exp, override = FALSE, kerpk = 1, fitmaxiter = 50,
  gam = 0.6, scl.factor = 0.1, tau = 2, maxwdth = 200)
}
\arguments{
\item{data.exp}{dataframe with experimental data (format?)}

\item{override}{force re-running of peak analysis}

\item{kerpk}{number of kernels per peak (passed to Ramanpk())}

\item{fitmaxiter}{number of max iterations while attempting fit (passed to Ramanpk())}

\item{gam}{gam (passed to Ramanpk())}

\item{scl.factor}{sclerosis factor (passed to Ramanpk())}

\item{tau}{tau (passed to Ramanpk())}

\item{maxwdth}{peak max width (passed to Ramanpk())}
}
\value{
a list of 4 elements (lists or dataframes)
}
\description{
This is the actual workhorse that does the peak fitting for each spectra.
Note: this function only works for a single spectra and does not handle
      saving the output and so on. Always use one of the wrapper functions.
}
