% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfitting.R
\name{RamanWrapper}
\alias{RamanWrapper}
\title{Use this wrapper function for calling Ramanpk}
\usage{
RamanWrapper(
  data.exp,
  run,
  override = FALSE,
  kerpk = 1,
  fitmaxiter = 50,
  gam = 0.6,
  scl.factor = 0.1,
  tau = 2,
  maxwdth = 200,
  jobfile = ""
)
}
\arguments{
\item{data.exp}{dataframe with experimental data (format?)}

\item{run}{index}

\item{override}{force re-running of peak analysis}

\item{kerpk}{number of kernels per peak (passed to Ramanpk())}

\item{fitmaxiter}{number of max iterations while attempting fit (passed to Ramanpk())}

\item{gam}{gam (passed to Ramanpk())}

\item{scl.factor}{sclerosis factor (passed to Ramanpk())}

\item{tau}{tau (passed to Ramanpk())}

\item{maxwdth}{peak max width (passed to Ramanpk())}

\item{jobfile}{path to datafile saved to disk, defaults to ./raman-peak-data.rda}
}
\value{
a Ramanpk dataframe
}
\description{
I wanted to run the \pkg{diffractometry} peak-analysis code (which can be both
time-consuming and prone to stray errors halting compilation) without the need
to re-run the peak analysis every time the document is re-compiled.
For that purpose, I created this wrapper function.
Its job is to call \code{\link{Ramanpk}} only if necessary (i.e., peak analysis
has not already been performed). That is made possible by saving the results
of a successful analysis to a file on disk (in the current working directory).
Also includes an override option to force re-running of peak analysis.
}
