% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-xtable.R
\name{ResetTabAttributes}
\alias{ResetTabAttributes}
\title{Reset xtable attributes for printing as LaTeX}
\usage{
ResetTabAttributes(table.object)
}
\arguments{
\item{table.object}{xtable object}
}
\value{
xtable object
}
\description{
Warning: this function needs the following variables to exist in the global namespace:
- tp
- obs
(this should be changed!). But for now, watch out.
}
