% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfitting.R
\name{SiliconWrapper}
\alias{SiliconWrapper}
\title{Use this wrapper function for silicon spectra}
\usage{
SiliconWrapper(data.exp, run, override = FALSE, kerpk = 1,
  fitmaxiter = 50, gam = 0.6, scl.factor = 0.1)
}
\arguments{
\item{data.exp}{dataframe with experimental data (format?)}

\item{run}{index}

\item{override}{force re-running of peak analysis}

\item{kerpk}{number of kernels per peak}

\item{fitmaxiter}{number of max iterations while attempting fit (diffractometry option)}

\item{gam}{gam (diffractometry option)}

\item{scl.factor}{sclerosis factor (diffractometry option)}
}
\value{
a Ramanpk dataframe
}
\description{
This is a more restricted wrapper, specifically intended for
use with silicon reference spectra.
}

