% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{Raman2df}
\alias{Raman2df}
\title{Import Renishaw Raman spectra files (ASCII)}
\usage{
Raman2df(datafilename, sampleid = "")
}
\arguments{
\item{datafilename}{text string with full path to experimental file}

\item{sampleid}{sampleid (optional, automatically created if not provided)}
}
\value{
Tibble with the following columns (and possibly some attributes):
   $ sampleid        : char
   $ wavenum         : numeric
   $ counts          : numeric
   $ wnum.interp     : interpolated wavenumber vector (equidistant)
   $ cts.interp      : interpolated counts vector
}
\description{
Reads Raman (ASCII) datafile and returns a dataframe with the original data,
as well as interpolated wavenumber and counts values
(interpolated data is evenly spaced along x-axis, required for later peak fitting
with diffractometry package).
}
